requirements.txt
flask
openai
fal-client
python-dotenv
asyncio
requests